refactor search service
