+++ packages/web-crawler/src/index.ts
@@ -1,2 +1,3 @@
+export type { CrawlImplType } from './crawImpl';
 export { Crawler } from './crawler';
 export * from './type';
