+++ src/server/routers/tools/search.ts
@@ -1,14 +1,9 @@
-import { Crawler } from '@lobechat/web-crawler';
-import { TRPCError } from '@trpc/server';
-import pMap from 'p-map';
 import { z } from 'zod';
 
-import { toolsEnv } from '@/config/tools';
 import { isServerMode } from '@/const/version';
 import { passwordProcedure } from '@/libs/trpc/edge';
 import { authedProcedure, router } from '@/libs/trpc/lambda';
-import { SearXNGClient } from '@/server/modules/SearXNG';
-import { SEARCH_SEARXNG_NOT_CONFIG } from '@/types/tool/search';
+import { searchService } from '@/server/services/search';
 
 // TODO: password procedure 未来的处理方式可能要思考下
 const searchProcedure = isServerMode ? authedProcedure : passwordProcedure;
@@ -22,24 +17,7 @@ export const searchRouter = router({
       }),
     )
     .mutation(async ({ input }) => {
-      const envString = toolsEnv.CRAWLER_IMPLS || '';
-
-      // 处理全角逗号和多余空格
-      let envValue = envString.replaceAll('，', ',').trim();
-
-      const impls = envValue.split(',').filter(Boolean);
-
-      const crawler = new Crawler({ impls });
-
-      const results = await pMap(
-        input.urls,
-        async (url) => {
-          return await crawler.crawl({ impls: input.impls, url });
-        },
-        { concurrency: 3 },
-      );
-
-      return { results };
+      return searchService.crawlPages(input);
     }),
 
   query: searchProcedure
@@ -56,25 +34,6 @@ export const searchRouter = router({
       }),
     )
     .query(async ({ input }) => {
-      if (!toolsEnv.SEARXNG_URL) {
-        throw new TRPCError({ code: 'NOT_IMPLEMENTED', message: SEARCH_SEARXNG_NOT_CONFIG });
-      }
-
-      const client = new SearXNGClient(toolsEnv.SEARXNG_URL);
-
-      try {
-        return await client.search(input.query, {
-          categories: input.optionalParams?.searchCategories,
-          engines: input.optionalParams?.searchEngines,
-          time_range: input.optionalParams?.searchTimeRange,
-        });
-      } catch (e) {
-        console.error(e);
-
-        throw new TRPCError({
-          code: 'SERVICE_UNAVAILABLE',
-          message: (e as Error).message,
-        });
-      }
+      return await searchService.query(input.query, input.optionalParams);
     }),
 });
